+++ packages/ui/src/main.ts
@@ -1,12 +1,16 @@
 import './components/vuestic-sass/resources/_css-variables.scss'
+
 export { VuesticPlugin } from './components/vuestic-plugin'
-export { useColor } from './services/color-config/ColorMixin'
+export { useColors, useColor } from './services/color-config/color-config'
 export { useGlobalConfig } from './services/GlobalConfigPlugin'
-export { useColors } from './services/color-config/color-config'
 export {
-  IconMixin,
   useIcons,
   createIconsConfig,
   VuesticIconFonts,
   VuesticIconAliases,
 } from './services/icon-config/icon-config'
+
+export type { GlobalConfig, GlobalConfigUpdater } from './services/GlobalConfigPlugin'
+export type { ComponentConfig } from './services/component-config/component-config'
+export type { IconsConfig, IconConfig } from './services/icon-config/icon-config'
+export type { ColorConfig } from './services/color-config/color-config'
