+++ packages/ui/src/services/color-config/ColorMixin.ts
@@ -1,38 +1,6 @@
 import { prop, Vue, Options, setup } from 'vue-class-component'
 
-import { useColors, DEFAULT_COLOR } from './color-config'
-
-/**
- * Check if color is valid css color
- * Taken from https://stackoverflow.com/a/56266358/5783475
- * @param strColor
- */
-export const isCssColor = (strColor: string): boolean => {
-  const s = new Option().style
-  s.color = strColor
-  return s.color !== ''
-}
-
-export const useColor = () => {
-  const { getColors = () => ({}) } = useColors() as any || {}
-  const colors = getColors() || {}
-
-  return (prop?: string, defaultColor: string = DEFAULT_COLOR): string => {
-    if (!prop) {
-      prop = defaultColor
-    }
-
-    if (colors[prop]) {
-      return colors[prop]
-    }
-
-    if (isCssColor(prop)) {
-      return prop
-    }
-
-    return defaultColor
-  }
-}
+import { useColor, useColors } from './color-config'
 
 class Props {
   color = prop<string>({
@@ -65,3 +33,7 @@ class ColorMixin extends Vue.with(Props) {
 }
 
 export default ColorMixin
+export {
+  useColors,
+  useColor,
+}
