+++ packages/ui/src/services/color-config/color-config.ts
@@ -25,3 +25,35 @@ export const useColors = () => {
     getColors,
   }
 }
+
+/**
+ * Check if color is valid css color
+ * Taken from https://stackoverflow.com/a/56266358/5783475
+ * @param strColor
+ */
+export const isCssColor = (strColor: string): boolean => {
+  const s = new Option().style
+  s.color = strColor
+  return s.color !== ''
+}
+
+export const useColor = () => {
+  const { getColors = () => ({}) } = useColors() as any || {}
+  const colors = getColors() || {}
+
+  return (prop?: string, defaultColor: string = DEFAULT_COLOR): string => {
+    if (!prop) {
+      prop = defaultColor
+    }
+
+    if (colors[prop]) {
+      return colors[prop]
+    }
+
+    if (isCssColor(prop)) {
+      return prop
+    }
+
+    return defaultColor
+  }
+}
