+++ packages/ui/src/services/icon-config/icon-config.ts
@@ -2,3 +2,4 @@ export { IconMixin } from './IconMixin'
 export { useIcons } from './setup'
 export { createIconsConfig } from './helpers'
 export { VuesticIconAliases, VuesticIconFonts } from './presets'
+export type { IconsConfig, IconConfig } from './types'
